refactor: restful한 서비스를 제공하기 위해 vote-list로 수정
