+++ apps/web/components/vote-list/VoteItem.tsx
